--- Tracker-ios/TrackerChoosingScreen/Controller/TrackerChoosingScreenController.swift
@@ -1,51 +1,0 @@
-import UIKit
-
-// MARK: - TrackerChoosingScreenController
-final class TrackerChoosingScreenController: UIViewController {
-
-    // MARK: - Properties and Initializers
-    override var preferredStatusBarStyle: UIStatusBarStyle {
-        return .darkContent
-    }
-
-    private let trackerChoosingScreenView = TrackerChoosingScreenView()
-
-    // MARK: - Life Cycle
-    override func viewDidLoad() {
-        view.addSubview(trackerChoosingScreenView)
-        view.backgroundColor = .ypWhite
-        setupConstraints()
-        trackerChoosingScreenView.delegate = self
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        if let topController = UIApplication.shared.windows.filter({ $0.isKeyWindow }).first?.rootViewController {
-            let destinationViewController = topController.children.first?.children.first as? TrackersScreenController
-            destinationViewController?.updateCollectionView()
-        }
-    }
-}
-
-// MARK: - Helpers
-extension TrackerChoosingScreenController {
-
-    private func setupConstraints() {
-        NSLayoutConstraint.activate([
-            trackerChoosingScreenView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            trackerChoosingScreenView.topAnchor.constraint(equalTo: view.topAnchor),
-            trackerChoosingScreenView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            trackerChoosingScreenView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-        ])
-    }
-}
-
-// MARK: - TrackerChoosingScreenViewDelegate
-extension TrackerChoosingScreenController: TrackerChoosingScreenViewDelegate {
-    func proceedToHabit() {
-        present(HabitCreationScreenController(), animated: true)
-    }
-
-    func proceedToEvent() {
-        present(HabitCreationScreenController(isNonRegularEvent: true), animated: true)
-    }
-}
